+++ OpenSim/Region/CoreModules/World/Terrain/PaintBrushes/LowerSphere.cs
@@ -50,10 +50,10 @@ namespace OpenSim.Region.CoreModules.World.Terrain.PaintBrushes
 
                     // Calculate a cos-sphere and add it to the heighmap
                     double r = Math.Sqrt(dx2 + (y - ry) * (y - ry));
-                    double distancefactor = Math.Cos(r * size);
+                    float distancefactor = (float)Math.Cos(r * size);
                     if (distancefactor > 0.0)
                     {
-                        double newz = map[x, y] - distancefactor * strength;
+                        float newz = map[x, y] - distancefactor * strength;
                         if (newz <= 0f)
                             map[x, y] = 0f;
                         else
